[FIX] hr_holidays: remove double-inheritance of hr.employee.base
Before this commit, two inheritance of hr.employee.base were done inside
the hr_holidays module. This commit merges both inheritance in the
hr_employee_base.py file

X-original-commit: 6b5cfca29413acd237d691dd33a5f76adef40412
